Added rotation to the pull events method. Still trying to figure out why certain pieces don't go all the way left

+++ Tetris/Tetris/TetroNMirror.cpp
@@ -13,7 +13,7 @@ TetroNMirror::~TetroNMirror()
 
 int TetroNMirror::getPiece(int pX, int pY)
 {
-	return (piece[actualRotation][pX][pY]);
+	return (piece[static_cast<int>(actualRotation)][pX][pY]);
 }
 
 void TetroNMirror::rotateRight()
